Addresses the latest review
